BL-572 revampled RenderTableContainer to harness async fetching of data. Major refactoring to make it readable, efficient, and to not have this particular component fetching the data (which should be shared between display components. Updated columns.js which is used for defining data columns to disinclude race/ethnicity, per the July 14 conversation with the stakeholder.
